requirements.txt
ccxt
pandas
numpy
requests
gspread
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
boto3
pandas-ta
ta